+++ movies.h
@@ -13,4 +13,4 @@
 #include <map>
 using namespace std;
 
-void print_alphabetical(map<string, int> mbin);
+void print_alphabetical(map<string, double> mbin);
